pyproject.toml
[tool.poetry]
name = "aiogram_template"
version = "0.0.6"
description = "A template written in aiogram"
authors = ["fonco <foncoofficial@gmail.com>"]
repository = "https://github.com/fonco/aiogram_template"
maintainers = [
    "F0rzend github.com/f0rzend",
    "comictomcat github.com/comictomcat",
]


[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^2.12"
PyYAML = "^5.4.1"
trafaret = "^2.1.0"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"
flake8 = "^3.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
